requirements.txt
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
fpdf2==2.7.9
python-dotenv==1.0.1
pytest==8.2.0